NSInvocation+RACTypeParsing: fix handling of typedef encodings.
Initially I had a method that received an `NSArray` argument. In another
place in code I added a `rac_signalForSelector:` for it. When I replaced
the `NSArray` with a `typedef`, and tried to access the its value sent
from the `rac_signalForSelector:`, an exception "`[NSConcreteValue
rac_sequence]` undefined signal to selector" was thrown.